Get rid of trailing slashes in Netlify
Netlify behavior with trailing slashes is explained at
https://answers.netlify.com/t/support-guide-how-can-i-alter-trailing-slash-behaviour-in-my-urls-will-enabling-pretty-urls-help/31191

At the moment the URLs are /llista/ and /blog/ because we are generating
the pages /llista/index.html and /blob/index.html. To get rid of the trailing
slash we need to instead generate /llista.html and /blog.html with the
format 'file'.

build.format docs: https://docs.astro.build/en/reference/configuration-reference/#buildformat

NavLink needs to be changed to continue being highlighted because pathname now is '/blog/2.html'.